Improve game fetching
