+++ newIDE/app/src/Export/Builds/BuildsDialog.js
@@ -6,54 +6,26 @@ import Dialog from '../../UI/Dialog';
 import HelpButton from '../../UI/HelpButton';
 import FlatButton from '../../UI/FlatButton';
 import Builds from '.';
-import AuthenticatedUserContext, {
-  type AuthenticatedUser,
-} from '../../Profile/AuthenticatedUserContext';
+import { type AuthenticatedUser } from '../../Profile/AuthenticatedUserContext';
 import useForceUpdate from '../../Utils/UseForceUpdate';
-import { getGame, type Game } from '../../Utils/GDevelopServices/Game';
+import { type Game } from '../../Utils/GDevelopServices/Game';
 
 type Props = {|
   authenticatedUser: AuthenticatedUser,
-  gameId: string,
+  game: Game,
   open: boolean,
   onClose: () => void,
+  onGameUpdated: () => void,
 |};
 
-const BuildsDialog = ({ authenticatedUser, gameId, open, onClose }: Props) => {
+const BuildsDialog = ({
+  authenticatedUser,
+  game,
+  open,
+  onClose,
+  onGameUpdated,
+}: Props) => {
   const forceUpdate = useForceUpdate();
-  const { getAuthorizationHeader, profile } = React.useContext(
-    AuthenticatedUserContext
-  );
-  const [game, setGame] = React.useState<?Game>(null);
-  const [isGameLoading, setIsGameLoading] = React.useState<boolean>(false);
-
-  const loadGame = React.useCallback(
-    async () => {
-      if (!profile || !gameId) return;
-
-      const { id } = profile;
-      try {
-        setIsGameLoading(true);
-        const game = await getGame(getAuthorizationHeader, id, gameId);
-        setGame(game);
-        setIsGameLoading(false);
-      } catch (err) {
-        setIsGameLoading(false);
-        console.error('Unable to load the game', err);
-      }
-    },
-    [gameId, getAuthorizationHeader, profile]
-  );
-
-  React.useEffect(
-    () => {
-      if (open) {
-        loadGame();
-      }
-    },
-    [loadGame, open]
-  );
-
   if (!open) return null;
 
   return (
@@ -80,8 +52,7 @@ const BuildsDialog = ({ authenticatedUser, gameId, open, onClose }: Props) => {
         onBuildsUpdated={forceUpdate}
         authenticatedUser={authenticatedUser}
         game={game}
-        onGameUpdated={setGame}
-        isGameLoading={isGameLoading}
+        onGameUpdated={onGameUpdated}
       />
     </Dialog>
   );
