+++ newIDE/app/src/Export/ExportDialog/index.js
@@ -12,6 +12,7 @@ import { useOnlineStatus } from '../../Utils/OnlineStatus';
 import AlertMessage from '../../UI/AlertMessage';
 import { Tab, Tabs } from '../../UI/Tabs';
 import ExportHome from './ExportHome';
+import { getGame, type Game } from '../../Utils/GDevelopServices/Game';
 
 const styles = {
   icon: { width: 40, height: 40 },
@@ -76,9 +77,42 @@ const ExportDialog = ({
   const [chosenExporterKey, setChosenExporterKey] = React.useState<ExporterKey>(
     'onlinewebexport'
   );
+  const [game, setGame] = React.useState<?Game>(null);
   const authenticatedUser = React.useContext(AuthenticatedUserContext);
+  const { getAuthorizationHeader, profile } = React.useContext(
+    AuthenticatedUserContext
+  );
   const onlineStatus = useOnlineStatus();
   const cantExportBecauseOffline = !!allExportersRequireOnline && !onlineStatus;
+
+  const loadGame = React.useCallback(
+    async () => {
+      if (!profile || !project) return;
+
+      const { id } = profile;
+      try {
+        const game = await getGame(
+          getAuthorizationHeader,
+          id,
+          project.getProjectUuid()
+        );
+        setGame(game);
+      } catch (err) {
+        console.error('Unable to load the game', err);
+      }
+    },
+    [project, getAuthorizationHeader, profile]
+  );
+
+  React.useEffect(
+    () => {
+      // Load game only once.
+      if (!game) {
+        loadGame();
+      }
+    },
+    [loadGame, game]
+  );
 
   if (!project) return null;
   const exporters = [
@@ -131,7 +165,7 @@ const ExportDialog = ({
           key="builds"
           label={<Trans>See this game builds</Trans>}
           onClick={() => setBuildsDialogOpen(true)}
-          disabled={isNavigationDisabled}
+          disabled={isNavigationDisabled || !game}
         />,
       ]}
       open
@@ -156,6 +190,7 @@ const ExportDialog = ({
           authenticatedUser={authenticatedUser}
           isNavigationDisabled={isNavigationDisabled}
           setIsNavigationDisabled={setIsNavigationDisabled}
+          onGameUpdated={setGame}
         />
       )}
       {chosenExporterSection === 'automated' && (
@@ -191,15 +226,19 @@ const ExportDialog = ({
             authenticatedUser={authenticatedUser}
             key={chosenExporterKey}
             setIsNavigationDisabled={setIsNavigationDisabled}
+            onGameUpdated={setGame}
           />
         </div>
       )}
-      <BuildsDialog
-        open={buildsDialogOpen}
-        onClose={() => setBuildsDialogOpen(false)}
-        authenticatedUser={authenticatedUser}
-        gameId={project.getProjectUuid()}
-      />
+      {game && (
+        <BuildsDialog
+          open={buildsDialogOpen}
+          onClose={() => setBuildsDialogOpen(false)}
+          authenticatedUser={authenticatedUser}
+          game={game}
+          onGameUpdated={setGame}
+        />
+      )}
     </Dialog>
   );
 };
